Statically generates radars from the md files
Thinking about what would be the best way of collaborating on the radar, github
and PRs seemed like a natural match as they would give visibility and
potentatially start conversations around any changes.

Having a single CSV file could be tricky, though, because of conflicts from
different branches all editing the same file so I thought maybe an easier way
would be to use the filesystem: add a file if you want to add an entry, etc.

So this is a MVP: I used a folder for each quadrant with a index.md that simply
contains a human readable name for the quadrant. Inside each folder there's a
md file for each node, named with the stage as a prefix, ie
`adopt-kubernetes.md`.  Inside this markdown file there's a title, an
indication if this is a New entry and free markdown text for descripition.

This generates a string with a CSV, on build, and so it produces a static radar
page with no external dependencies for data.  Let me know what you think, if
this could be a way to go or if it is too hacky, etc.